Backend/Urdu-OCR-APP/files/requirements.txt
fastapi
uvicorn
pdf2image
Pillow
google-genai
python-multipart
